requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.3.2
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-grappelli==4.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
